requirements.txt
alembic==1.18.4
bcrypt==4.0.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.20.1
packaging==26.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.48
stripe==7.5.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
